% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_na_rows.R
\name{all_na_rows}
\alias{all_na_rows}
\alias{all_na_columns}
\title{Find rows and columns in which all values are NA's}
\usage{
all_na_rows(x)

all_na_columns(x)
}
\arguments{
\item{x}{A data frame or a matrix.}
}
\value{
A logical vector with rows/columns in which all values are
\code{NA} indicated as \code{TRUE}.
}
\description{
Find rows and columns in which all values are NA's
}
\examples{

 x <- matrix(c(NA, 1, 2, NA, 3, 4, NA, NA, NA), nrow = 3)
 x
 all_na_rows(x)

 all_na_columns(x)


}
