% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_.R
\name{eval_}
\alias{eval_}
\title{[!] Parse a string and evaluate expression}
\usage{
eval_(X, envir = parent.frame())
}
\arguments{
\item{X}{A string to be evaluated as an expression.}

\item{envir}{the \code{\link{environment}} in which \code{expr} is to
    be evaluated.  May also be \code{NULL}, a list, a data frame,
    a pairlist or an integer as specified to \code{\link{sys.call}}.}
}
\value{
Evaluated expression
}
\description{
A wrapper function to parse a string and evaluate it as an expression.
\code{eval_(X)} is the same as \code{eval(parse(text = X)}.
}
\examples{
if (any(ls() == "A")) rm(A)
any(ls() == "A")
#> [1] FALSE

eval_("A <- 3")
any(ls() == "A")
#> TRUE

}
\seealso{
Evaluate an (unevaluated) expression (\link[base]{eval}),
         Parse expressions (\link[base]{parse})
}

