% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fCap.R
\name{fCap}
\alias{fCap}
\alias{make_firstCapitals}
\title{[+] Convert first letters of all words to capitals}
\usage{
fCap(x)
}
\arguments{
\item{x}{Either a string or a vector of strings.}
}
\value{
The same string as input with all words starting in capital letters.
}
\description{
Convert first letters of all words in a string to capitals.
}
\examples{

fCap('laa laa laa')
##[1] "Laa Laa Laa"

}
\seealso{
Other \pkg{spMisc} utilities: 
\code{\link{bru}()},
\code{\link{clc}()},
\code{\link{clear}()},
\code{\link{isFALSE}()},
\code{\link{list_AddRm}()},
\code{\link{make.filenames}()},
\code{\link{open_wd}()},
\code{\link{printDuration}()},
\code{\link{regexp2df}()},
\code{\link{st01}()}
}
\concept{\pkg{spMisc} utilities}
