% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_to_regexp.R
\name{fixed_to_regexp}
\alias{fixed_to_regexp}
\title{Convert fixed expression to regular expression}
\usage{
fixed_to_regexp(str)
}
\arguments{
\item{str}{A string}
}
\value{
The string rewritten as regular expression.
}
\description{
A back slash (\code{\ }) is added before special symbols
(\code{. \ | ( ) [ \{ \}^ $ * + ?}) in a string \code{str}
to interpret them as "fixed" in a regular expression.
Useful in functions, which accept regular expressions but do not have
option \code{fixed = TRUE}.
}
\examples{
fixed_to_regexp("\\\\")
  ## "\\\\"

fixed_to_regexp("||")
fixed_to_regexp(".\\\\|(){}^$*+?[]")

fixed_to_regexp("(A|B)")
 ## "\\(A\\|B\\)"

fixed_to_regexp("Hello!") # nothing is changed
  ## "Hello!"

}
