% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullfile.R
\name{fullfile}
\alias{fullfile}
\title{Build full file name from parts}
\usage{
fullfile(..., ext = "", fsep = .Platform$file.sep)
}
\arguments{
\item{...}{character vectors.  Long vectors are not supported.}

\item{ext}{- extension of file name (e.g. ".txt" or "txt")}

\item{fsep}{the path separator to use (assumed to be ASCII).}
}
\value{
A character vector of the arguments concatenated term-by-term and
separated by \code{fsep} if all arguments have positive length;
otherwise, an empty character vector (unlike \code{\link[base]{paste}}).
All slashes are converted to \code{fsep}. Double slashes are removed.
}
\description{
Build full file name from parts and remove repeated slashes.
Similar to MATLAB's function "fulfile"
}
\examples{

fullfile("a","b","c")
 #> [1] "a/b/c"

fullfile("a/","/b/","/c/")
 #> [1] "a/b/c"

fullfile("a","b\\\\","c", ext = "txt")
 #> [1] "a/b/c.txt"

fullfile("a","b\\\\","c", ext = "txt")
 #> [1] "a/b/c.txt"

}
