% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_enable_favicons.R
\name{html_enable_favicons}
\alias{html_enable_favicons}
\title{Add code to HTML head that enables display of favicons}
\usage{
html_enable_favicons(main_dir = "./docs", favicon_path = "favicon.png")
}
\arguments{
\item{main_dir}{(string(1)) \cr root directory of your website.}

\item{favicon_path}{(string(1)) \cr relative path to favicon
(from the \code{main_dir}).}
}
\value{
A vector with messages if procedure suceeded
}
\description{
Function adds a line of code to the last line of head of all \code{.html}
and \code{.htm} files in the indicated main directory and its subdirectories.
\bold{NOTE:} function doos not check if the favicon exists in the directory.
}
\examples{
# html_enable_favicons("./docs", "favicon.ico")
}
\seealso{
html_add_line_to_head
}
\author{
Vilmantas Gegzna
}
