% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indMatrix.R
\name{indMatrix}
\alias{indMatrix}
\title{Create a matrix filled with its element indices}
\usage{
indMatrix(x)
}
\arguments{
\item{x}{A matrix.}
}
\description{
Genereta a matrix of the same size as provided matrix \code{x}, where each
element represents its index.
}
\examples{

 indMatrix(matrix(NA, 2, 5))
 indMatrix(matrix(NA, 5, 2))

}
\author{
Vilmantas Gegzna
}
\seealso{
\link[base]{row},\link[base]{col}

Other matrix operations in \pkg{spMisc}: \code{\link{corr_vec2mat}},
  \code{\link{nTri2nDiag}}, \code{\link{revalueMatrix}},
  \code{\link{sort_matrix_elements}},
  \code{\link{which.in}}, \code{\link{which.max.all}},
  \code{\link{which.max.perRow}}
}

