% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_operators.R
\name{Insert if NULL}
\alias{Insert if NULL}
\alias{\%if.NULL\%}
\title{Infix operator to insert default value if result is NULL}
\source{
Inspired by "Inxix functions" in
\href{http://adv-r.had.co.nz/Functions.html#function-arguments}{Advanced R}
by Hadley Wickham.
}
\usage{
a \%if.NULL\% b
}
\arguments{
\item{a}{result of a function.}

\item{b}{default value is \code{a} is \code{NULL}.}
}
\value{
Either result of expresion \code{a} if it is not \code{NULL}, or
epression \code{b} otherwise.
}
\description{
The function is useful as a way of providing a default value
 in case the output of another function is \code{NULL}.
}
\examples{
a1 <- 1
a2 <- NULL
b  <- "Default value"

 a1 \%if.NULL\% b
#> [1] 1

 a2 \%if.NULL\% b
#> "Default value"

}
\seealso{
\code{\link[spAddins]{insertIfNULL_Addin}}
}

