% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isFALSE.R
\name{isFALSE}
\alias{ifFALSE}
\alias{isFALSE}
\title{Is an expresion equal to FALSE?}
\usage{
isFALSE(condition)

ifFALSE(condition, is_false, otherwise_)
}
\arguments{
\item{condition}{Either a logical condition or \code{NULL}.}

\item{is_false}{Result if condition is equal to  \code{FALSE}.}

\item{otherwise_}{Result if condition is not equal to \code{FALSE}.}
}
\description{
If condition is equal to \code{FALSE}, returns \code{TRUE},
otherwise (i.e. if \code{NULL} or \code{TRUE}) returns \code{FALSE}.
}
\author{
Vilmantas Gegzna
}
\seealso{
\code{\link[base]{isTRUE}}

Other \pkg{spMisc} utilities: \code{\link{bru}},
  \code{\link{fCap}}, \code{\link{list_AddRm}},
  \code{\link{open_wd}}
}

