% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex2rmd.R
\name{latex2rmd}
\alias{latex2rmd}
\title{Convert Rnw to Rmd}
\usage{
latex2rmd(input, output)
}
\arguments{
\item{output}{}
}
\description{
Does most of the job converting Sweave .Rnw files to R markdown .Rmd format.
 Not perfect, but good enough.
}
\examples{

# // none //
}
