% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_AddRm.R
\name{list_AddRm}
\alias{list_AddRm}
\title{[.] Find and display added and removed names}
\usage{
list_AddRm(initial, final)
}
\arguments{
\item{initial}{A charecter vector with initial names.}

\item{final}{A charecter vector with final names.}
}
\value{
A list with added and removed names
}
\description{
[.] Find and display added and removed names
}
\author{
Vilmantas Gegzna
}
\seealso{
Other \pkg{spMisc} utilities: \code{\link{bru}},
  \code{\link{fCap}}, \code{\link{isFALSE}},
  \code{\link{open_wd}}
}

