% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.filenames.R
\name{make.filenames}
\alias{make.filenames}
\title{[!+] Convert string into Windows compatible filename}
\usage{
make.filenames(s, replacement = "_", allow.space = TRUE)
}
\arguments{
\item{s}{A string.}

\item{replacement}{A symbol to replce unallowed symbols with.
\code{replacement} must be an allowed symbol.}

\item{allow.space}{Logical. If \code{FALSE}, space will be
repalaced with \code{replacement}. Default is \code{TRUE}.}
}
\value{
A string without special symbols.
}
\description{
Function replaces special symbols (\code{\"  / : * ?  < > |}) in
a string \code{s} with \code{replacement} (by default it is underscore
 \code{_}). After replacement the  string can be used as a file name
 in Windows.
}
\examples{
 s <- '\\\\ / : * ? " < > |'
 make.filenames(s)
 #> "_ _ _ _ _ _ _ _ _"

 make.filenames(s, allow.space = FALSE)
 #> "_________________"


 s2 <- "/Hello?"
 make.filenames(s2)
 #> "_Hello_"

 make.filenames(s2, replacement = "-")
 #> "-Hello-"

}
\seealso{
Other \pkg{spMisc} utilities: \code{\link{bru}},
  \code{\link{clc}}, \code{\link{clear}},
  \code{\link{fCap}}, \code{\link{isFALSE}},
  \code{\link{list_AddRm}}, \code{\link{open_wd}},
  \code{\link{printDuration}}, \code{\link{regexp2df}},
  \code{\link{st01}}
}

