% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na.replace.R
\name{na.replace}
\alias{na.replace}
\title{Mask NA values with other symbol}
\usage{
na.replace(obj, replacement = " ")
}
\arguments{
\item{obj}{an objecta, that may have \code{NA} values}

\item{replacement}{a replacement for \code{NA} values. default is spece (\code{" "}).}
}
\value{
the same object with NA values replaced.
}
\description{
Replace (i.e. mask) NA values with other symbol
}
\examples{

library(pander)
library(magrittr)

summary(CO2[,3:4])  \%>\% pander()
summary(CO2[,3:4])  \%>\% na.replace() \%>\% pander()


}
