% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_wd.R
\name{open_wd}
\alias{open_wd}
\title{[+] Open curent working directory}
\usage{
open_wd()
}
\value{
No value, but informative error messages will be given if the
operation fails.
}
\description{
Function opens curent working directory.
}
\details{
This function is a wrapper for \code{shell.exec(getwd())}.
}
\examples{

\donttest{
\dontrun{
open_wd()
}}

}
\seealso{
\code{\link[base]{shell.exec}}

Other \pkg{spMisc} utilities: \code{\link{bru}},
  \code{\link{clear}}, \code{\link{fCap}},
  \code{\link{isFALSE}}, \code{\link{list_AddRm}},
  \code{\link{make.filenames}},
  \code{\link{printDuration}}, \code{\link{regexp2df}},
  \code{\link{st01}}
}

