% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgdown_find_missing_references.R
\name{pkgdown_find_missing_references}
\alias{pkgdown_find_missing_references}
\alias{print.missing_references}
\title{Find missing function references in "_pkgdown.yaml" file}
\usage{
pkgdown_find_missing_references()

\method{print}{missing_references}(x, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Vector of missing entries
}
\description{
Find missing function references in "_pkgdown.yaml" file
}
\examples{
pkgdown_find_missing_references()
}
