% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printDuration.R
\name{printDuration}
\alias{printDuration}
\alias{Duration}
\alias{time_elapsed}
\title{[+!] Evaluate how much time passed}
\usage{
printDuration(Start, Message = "Analysis completed in",
  returnString = FALSE, End = Sys.time())

Duration(Start, Message = "Analysis completed in", print = FALSE,
  End = Sys.time())

time_elapsed(start = stop("'start' is missing"), end = Sys.time())
}
\arguments{
\item{Start, start}{Moment of time which is treated as the beggining (object of class
\code{\link[=POSIXct-class]{POSIXct}}).}

\item{Message}{Message before time stamp that describes it. Default is
\code{"Duration of analysis:"}}

\item{returnString}{If \code{TRUE}, returns result as a string.
If \code{FALSE} (default), function \code{\link[pander]{pander}} prints the
result.}

\item{End, end}{Moment of time which is treated as the end (object of class
\code{\link[=POSIXct-class]{POSIXct}}).}

\item{print}{(logical) should the result be printed using package \pkg{pander}?}
}
\value{
Text indicating how much time has passed (either printed in console or as a string).
}
\description{
Evaluate and print difference between moment of time at which time
interval began (usually captured by function \code{\link[base]{Sys.time}} some time ago) and
the present (i.e., moment when function \code{printDuration} is called).
}
\examples{
Start <-  Sys.time()

Start
## [1] "2016-02-12 16:15:09 UTC"

class(Start)
## [1] "POSIXct" "POSIXt"

printDuration(Start)
## Duration of analysis: 23.3 secs

printDuration(Start,"From start till now")
## From start till now 39.2 secs

my_duration <- printDuration(Start)
my_duration
## NULL

my_duration <- printDuration(Start, returnString = TRUE)
my_duration
## [1] "Analysis completed in 2.4 mins"

}
\seealso{
\code{\link[base]{difftime}}

Other \pkg{spMisc} utilities: \code{\link{bru}},
  \code{\link{clc}}, \code{\link{clear}},
  \code{\link{fCap}}, \code{\link{isFALSE}},
  \code{\link{list_AddRm}}, \code{\link{make.filenames}},
  \code{\link{open_wd}}, \code{\link{regexp2df}},
  \code{\link{st01}}
}
\author{
Vilmantas Gegzna
}
\concept{\pkg{spMisc} utilities}
