% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbind_df_in_list.R
\name{rbind_df_in_list}
\alias{rbind_df_in_list}
\title{Rowbind dataframes, which are elements of a named list}
\usage{
rbind_df_in_list(x)
}
\arguments{
\item{x}{a named list, that contain dataframes that have columns with the same names.}
}
\value{
A data frame.
}
\description{
Rowbind dataframes, which are elements of a named list
}
\examples{

LIST <- split(chickwts, chickwts$feed)
new_DF <- rbind_df_in_list(LIST)

head(chickwts)

#    weight      feed
#  1    179 horsebean
#  2    160 horsebean
#  3    136 horsebean
#  4    227 horsebean
#  5    217 horsebean
#  6    168 horsebean


head(new_DF)

#    .group weight   feed
# 1 casein    368 casein
# 2 casein    390 casein
# 3 casein    379 casein
# 4 casein    260 casein
# 5 casein    404 casein
# 6 casein    318 casein

}
