% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regcapturedmatches.R
\name{regcapturedmatches}
\alias{regcapturedmatches}
\title{[.] Extract regular expression matches captured by names}
\source{
\href{https://gist.github.com/MrFlick/10413321}{regcapturedmatches.R} on gist.github.com
\href{http://stackoverflow.com/questions/33288075/from-matlab-to-r-capture-named-fields-with-regular-expressions-to-a-dataframe}{answer} on stackoverflow.com.
}
\usage{
regcapturedmatches(x, m)
}
\arguments{
\item{x}{(A list of) strings.}

\item{m}{Parsed data, a result from a regular expression function.}
}
\value{
A list with captured matches
}
\description{
regcapturedmatches.R: extracts captured matches from match data obtained
 by regexpr, gregexpr or regexec.
}
\examples{
# usage

x<-c("larry:35,M","alison:22,F","dave","lily:55,F")
m<-regexpr("(.*):(\\\\d+),([MF])", x, perl=TRUE)
spMisc:::regcapturedmatches(x,m)

}
\author{
\href{http://stackoverflow.com/users/2372064/mrflick?tab=profile}{MrFlick}
}
