% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_names.R
\name{remove_names}
\alias{remove_names}
\title{Remove names of an object}
\usage{
remove_names(x)
}
\arguments{
\item{x}{An object, which have attribute "names"}
}
\value{
The same object without names.
}
\description{
Remove names of an object
}
\examples{
obj <- c(a = 1)
obj

remove_names(obj)

}
