% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_mat.R
\name{rep_mat}
\alias{rep_mat}
\alias{rep_cols}
\alias{rep_rows}
\title{Repeat object column-wise and row-wise}
\usage{
rep_mat(obj, n = c(nrows, ncols), ncols = 1, nrows = 1)

rep_cols(obj, ncols = 1)

rep_rows(obj, nrows = 1)
}
\arguments{
\item{obj}{An object to be repeated (a matrix, a vector, ...).}

\item{n}{(vector of 2 integers | integer) a number of repetitions
row-wise and column-wise. If one integer may be indicated, it is
treated as \code{nrows}.}

\item{ncols}{(integer) a number of repetitions column-wise. Ignored if
second element of \code{n} is provided.}

\item{nrows}{(integer) a number of repetitions row-wise. Ignored if
\code{n} is provided.}
}
\value{
The same object, repeated column-wise and row-wise.
}
\description{
Repeat object column-wise and row-wise
}
\examples{
library(spMisc)

M <- matrix(1:4,2,2)

# These pairs of functions do the same:
     rep_mat(M, c(2,3))
     rep_mat(M, 2, 3)

     rep_mat(M, ncols = 2)
     rep_cols(M, 2)

     rep_mat(M, nrows = 2)
     rep_rows(M, 2)

}
