% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_DescOnDiag.R
\name{sort_DescOnDiag}
\alias{sort_DescOnDiag}
\alias{sort_colSums}
\alias{sort_maxValueInCol}
\alias{sort_maxValueInRow}
\alias{sort_rowSums}
\title{[!!!] Sort rows and columns of a matrix}
\usage{
sort_DescOnDiag(M)

sort_colSums(M)

sort_rowSums(M)

sort_maxValueInRow(M)

sort_maxValueInCol(M)
}
\arguments{
\item{M}{A matrix.}
}
\description{
[!!!] Sort rows and columns of a matrix
}
\examples{
library(spMisc)

M0 <- matrix(c(0, 11,8, 0,
               12,0, 0, 0,
               0, 0, 0, 9,
               0, 0, 9, 0),byrow=TRUE,nc=4)
M0 <- pkgmaker::addnames(M0)

M0
sort_DescOnDiag(M0)
sort_colSums(M0)
sort_rowSums(M0)
sort_maxValueInRow(M0)
sort_maxValueInCol(M0)



M1 <- sort_DescOnDiag(M0)
print(M0)
print(M1)



}
\author{
Vilmantas Gegzna
}
\seealso{
Other matrix operations in \pkg{spHelper}: \code{\link{nTri2nDiag}},
  \code{\link{newMatrix}}, \code{\link{which.in}},
  \code{\link{which.max.all}},
  \code{\link{which.max.perRow}}
}

