% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visNetwork_.R
\name{visNetwork_}
\alias{visNetwork_}
\alias{file2text}
\title{Wrapper for function visNetwork}
\usage{
visNetwork_(
  file,
  width = NULL,
  height = NULL,
  main = NULL,
  encoding = "UTF-8",
  ...
)

file2text(file, encoding = "UTF-8")
}
\arguments{
\item{file}{path to graphviz (.gv) file}

\item{width}{: Width (optional, defaults to automatic sizing)}

\item{height}{: Height (optional, defaults to automatic sizing)}

\item{main}{: For add a title. Character or a named list.
\itemize{
 \item{"text"}{ : Character. Title.}
 \item{"style"}{ : Optional. Character. HTML style of title. Default to 'font-family:Georgia, Times New Roman, Times, serif;font-weight:bold;font-size:20px;text-align:center;'.}
}}

\item{encoding}{encoding to be assumed for input strings.  It is
    used to mark character strings as known to be in
    Latin-1 or UTF-8: it is not used to re-encode the input.  To do the
    latter, specify the encoding as part of the connection \code{con} or
    via \code{\link[base]{options}(encoding=)}: see the examples.
    See also \sQuote{Details}.
  }

\item{...}{: Don't use.}
}
\description{
The result is the same as if function \code{visNetwork} was used.
}
