% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isFALSE.R
\name{isFALSE}
\alias{isFALSE}
\alias{ifFALSE}
\title{Is an expresion equal to FALSE?}
\usage{
isFALSE(condition)

ifFALSE(condition, is_false, otherwise_)
}
\arguments{
\item{condition}{Either a logical condition or \code{NULL}.}

\item{is_false}{Result if condition is equal to  \code{FALSE}.}

\item{otherwise_}{Result if condition is not equal to \code{FALSE}.}
}
\description{
If condition is equal to \code{FALSE}, returns \code{TRUE},
otherwise (i.e. if \code{NULL} or \code{TRUE}) returns \code{FALSE}.
}
\seealso{
\code{\link[base]{isTRUE}}

Other \pkg{spMisc} utilities: 
\code{\link{bru}()},
\code{\link{clc}()},
\code{\link{clear}()},
\code{\link{fCap}()},
\code{\link{list_AddRm}()},
\code{\link{make.filenames}()},
\code{\link{open_wd}()},
\code{\link{printDuration}()},
\code{\link{regexp2df}()},
\code{\link{st01}()}
}
\author{
Vilmantas Gegzna
}
\concept{\pkg{spMisc} utilities}
