% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which.max.all.R
\name{which.max.all}
\alias{which.max.all}
\alias{which.min.all}
\title{[!] Get indices of minima/maxima}
\usage{
which.max.all(x, ...)

which.min.all(x, useNames = FALSE, ...)
}
\arguments{
\item{x}{numeric (logical, integer or double) vector or an \R object
    for which the internal coercion to \code{\link{double}} works whose
    \code{\link{min}} or \code{\link{max}} is searched for.}

\item{...}{Further parameters for \code{\link[base]{which}}.}

\item{useNames}{logical indicating if the value of \code{arrayInd()}
   should have (non-null) dimnames at all.}
}
\description{
Determines the location, i.e., index of all minimum or maximum of a
numeric (or logical) vector.
}
\details{
Internaly functions \code{\link[base]{which}} and
\code{\link[base]{as.vector}} are used.
}
\examples{
x <- c(1,2,1,3,3,2,1)
which.max.all(x)
#>  4 5

which.min.all(x)
#>  1 3 7

}
\author{
Vilmantas Gegzna
}
\seealso{
Other matrix operations in \pkg{spHelper}: \code{\link{nTri2nDiag}},
  \code{\link{newMatrix}}, \code{\link{sort_DescOnDiag}},
  \code{\link{which.in}}, \code{\link{which.max.perRow}}
}

